fix TypeError: 'unicode' does not have the buffer interface and update for python 3 support
from __future__ import unicode_literals - breaks this module. It appears that pypureomapi leverage sockets
that are designed to only take byte strings. In python 2, strings are already byte strings and using unicode_literals
causes the TypeError. In python 3 strings are unicode by default and thus need to be encoded to byte strings.
In this module, string literals have been explicitly updated to be encoded into byte strings so as to handle both python 2 and 3.